requirements.txt
aiogram==3.13.1
apscheduler==3.10.4
httpx==0.27.2
python-dotenv==1.0.1
playwright==1.48.0
beautifulsoup4==4.12.3
lxml==5.2.1
pydantic==2.9.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
Jinja2==3.1.4
